fixed warning

+++ src/md_xtc.c
@@ -314,7 +314,7 @@ static bool xtc_decode_frame_data(struct md_trajectory_o* inst, const void* fram
             result = xtc_frame_coords(file, natoms, pos);
             if (result) {            
                 // nm -> Ångström
-                for (size_t i = 0; i < natoms; ++i) {
+                for (int i = 0; i < natoms; ++i) {
                     if (x) x[i] = pos[i][0] * 10.0f;
                     if (y) y[i] = pos[i][1] * 10.0f;
                     if (z) z[i] = pos[i][2] * 10.0f;
